fix: Rename `default` resume, letter and job-application files as `resume`, `letter` and `job-application` respectively.
