--- packages/job-applications/jobApplication.js
@@ -1,8 +1,0 @@
-const resume = require("@randy.tarampi/resume/resume");
-const letter = require("@randy.tarampi/letter/letter");
-
-module.exports = {
-    id: "default",
-    letter,
-    resume
-};
