+++ packages/letter/package.json
@@ -103,6 +103,6 @@
     "dist/styles.css",
     "index.client.js",
     "index.server.js",
-    "letter.json"
+    "src/letters/letter.json"
   ]
 }
