+++ packages/resume/gulpfile.js
@@ -90,7 +90,7 @@ gulp.task("resume:json", done => {
     const fs = require("fs");
     const config = require("config");
     const path = require("path");
-    const baseResumePath = path.join(__dirname, "./resume.json");
+    const baseResumePath = path.join(__dirname, "src/resumes/resume.json");
     const resume = require(baseResumePath);
     const {Person} = require("@randy.tarampi/js");
     const basics = Person.fromJSON(config.get("me.resume.basics")).toResume();
