+++ packages/resume/src/resumes/index.js
@@ -1,3 +1,3 @@
-import defaultResume from "../../resume";
+import defaultResume from "./resume.json";
 
 export default defaultResume;
